+++ service/cart/src/main/java/com/bappul/cart/adapter/pricing/response/PricingInternalResponse.java
@@ -1,6 +1,7 @@
-package com.bappul.cart.client.response;
+package com.bappul.cart.adapter.pricing.response;
 
 import java.math.BigDecimal;
+import java.util.List;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -11,8 +12,9 @@ import lombok.experimental.FieldDefaults;
 @Builder
 @AllArgsConstructor
 @FieldDefaults(level = AccessLevel.PRIVATE)
-public class OptionPerPrice {
-  Long optionValueId;
-  String optionName;
-  BigDecimal optionPrice;
+public class PricingInternalResponse {
+  Long storeId;
+  String storeName;
+  List<CartItemCalculateResponse> items;
+  BigDecimal totalPrice;
 }
